pyproject.toml
[tool.poetry]
name = "syncious"
version = "0.0.1"
description = ""
authors = ["WardPearce <wardpearce@pm.me>"]
readme = "README.md"
packages = [
    {include = "*", from = "syncious"},
    {include = "run.py"}
]

[tool.poetry.scripts]
server = "run:main"

[tool.poetry.dependencies]
python = "^3.12"
litestar = {extras = ["full"], version = "^2.8.0"}
aiohttp = "^3.9.3"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
tortoise-orm = {extras = ["asyncpg"], version = "^0.20.1"}
sqlalchemy = "^2.0.29"
uvicorn = "^0.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
